move `HideAsFalseAction` to shared utils module
Seems like a handy argparse action that can be used by other modules.
Suggested by @huddlej in review
